Update the start command from beta
Do it like it's done in mbently. Added the flag:
--add-opens java.base/java.util=ALL-UNNAMED:
This is a JVM option introduced in Java 9. It allows the
specified package (java.util from java.base) to be opened for
reflection by all unnamed modules, which can be useful for
compatibility with older code that relies on reflection.